pyproject.toml
[tool.poetry]
name = "reviews-api"
version = "0.1.0"
description = ""
authors = ["kirill43qw <default11user@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"
django = "^5.1"
python-dotenv = "^1.0.1"
django-filter = "^24.3"
django-environ = "^0.11.2"
django-ninja = "^1.3.0"
punq = "^0.7.0"
httpx = "^0.27.2"
python-slugify = "^8.0.4"


[tool.poetry.group.dev.dependencies]
faker = "^28.4.1"
factory-boy = "^3.3.1"
ipython = "^8.28.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = 'core.project.settings.local'
